feat: added skeleton on deposit page
