+++ src/components/Input/index.jsx
@@ -1,0 +1,1 @@
+export { Input } from './Input';
